requirements.txt
# Core dependencies
Flask==3.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# API and HTTP
openai==1.65.5
requests==2.32.3
urllib3==2.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10

# HTTP Libraries for OpenAI
httpx==0.28.1
httpcore==1.0.7
anyio==4.8.0
sniffio==1.3.1
h11==0.14.0

# Data validation and utilities
pydantic==2.10.6
pydantic-core==2.27.2
annotated-types==0.7.0
typing_extensions==4.12.2
tqdm==4.67.1
distro==1.9.0

# Environment variables
python-dotenv==1.0.1

# Note: This project uses PayPal REST API in sandbox mode only
